requirements.txt
Django==1.6.7
south==1.0.2
Pillow==3.4.2
42cc-pystyle
